// frontend/src/generated/apiv2/models/JsonWebKeyCurveName.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cryptocat API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const JsonWebKeyCurveName = {
    CurveNameP256: 'P-256',
    CurveNameP384: 'P-384',
    CurveNameP521: 'P-521'
} as const;
export type JsonWebKeyCurveName = typeof JsonWebKeyCurveName[keyof typeof JsonWebKeyCurveName];


export function JsonWebKeyCurveNameFromJSON(json: any): JsonWebKeyCurveName {
    return JsonWebKeyCurveNameFromJSONTyped(json, false);
}

export function JsonWebKeyCurveNameFromJSONTyped(json: any, ignoreDiscriminator: boolean): JsonWebKeyCurveName {
    return json as JsonWebKeyCurveName;
}

export function JsonWebKeyCurveNameToJSON(value?: JsonWebKeyCurveName | null): any {
    return value as any;
}